chore: Change Property Type From Long To Int

+++ data/night-study/src/commonMain/kotlin/com/b1nd/dodam/data/nightstudy/model/NightStudyStudent.kt
@@ -5,9 +5,9 @@ import com.b1nd.dodam.network.nightstudy.model.NightStudyStudentResponse
 data class NightStudyStudent(
     val id: Long,
     val name: String,
-    val grade: Long,
-    val room: Long,
-    val number: Long,
+    val grade: Int,
+    val room: Int,
+    val number: Int,
     val phone: String,
     val profileImage: String,
     val isBanned: Boolean
